third_party/blink/web_tests/dom/mutation-event-tests/fast/dom/getElementById-consistency-expected.txt

CONSOLE WARNING: Listener added for a synchronous 'DOMSubtreeModified' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
Test that DOMSubtreeModified listeners cannot cause inDocument to be incorrect

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.contains(document.getElementById('el')) is true
PASS document.getElementById('el') is null
PASS successfullyParsed is true

TEST COMPLETE
